Change float into int

+++ march_hardware/include/march_hardware/IMotionCube.h
@@ -55,7 +55,7 @@ public:
   void setControlWord(uint16 controlWord);
 
   void actuateRad(float targetRad);
-  void actuateCurrent(float targetCurrent);
+  void actuateCurrent(int targetCurrent);
 
   void actuateRadFixedSpeed(float targetRad, float radPerSec);
